updated cmdata and make_mat to deal with h5 files also cmdata to float

+++ tools/cmdata/src/cmdata/density.hpp
@@ -10,33 +10,33 @@
 namespace cmdata::density
 {
 
-void kernel_density_estimator(std::vector<double>::iterator x, const std::vector<double> &bins, const double mu, const double norm)
+void kernel_density_estimator(std::vector<float>::iterator x, const std::vector<float> &bins, const float mu, const float norm)
 {
-  double h = 0.01;
-  double from_x = std::max(mu - 2 * h, bins[0]);
-  double to_x = std::min(mu + 2 * h, bins.back());
-  auto is_geq_start = [&from_x](double i) { return i >= from_x; };
-  auto is_geq_end = [&to_x](double i) { return i > to_x; };
+  float h = 0.01;
+  float from_x = std::max(mu - 2 * h, bins[0]);
+  float to_x = std::min(mu + 2 * h, bins.back());
+  auto is_geq_start = [&from_x](float i) { return i >= from_x; };
+  auto is_geq_end = [&to_x](float i) { return i > to_x; };
   auto start = std::find_if(bins.begin(), bins.end(), is_geq_start);
   auto end = std::find_if(bins.begin(), bins.end(), is_geq_end);
   int from = std::distance(bins.begin(), start);
   int to = std::distance(bins.begin(), end);
-  double scale = norm / (0.73853587 * h * std::sqrt(2. * M_PI));
+  float scale = norm / (0.73853587 * h * std::sqrt(2. * M_PI));
   if (mu < h) scale *= 2.;
-  double shift = std::exp(-2.);
+  float shift = std::exp(-2.);
   for (int i = from; i < to; i++)
   {
-    double f = (mu - bins[i]) / h;
-    double kernel = std::exp(-0.5 * f * f);
+    float f = (mu - bins[i]) / h;
+    float kernel = std::exp(-0.5 * f * f);
     x[i] += scale * (kernel - shift);
   }
 }
 
 void intra_mol_routine( 
-  int i, std::size_t a_i, std::size_t a_j, double dx2, double weight, const std::vector<int> &mol_id_,
-  const std::vector<int> &natmol2_, const std::vector<double> &density_bins_,
-  const std::vector<double> &inv_num_mol_, std::vector<std::vector<std::mutex>> &frame_same_mutex_, 
-  std::vector<std::vector<std::vector<std::vector<double>>>> &intram_mat_density_
+  int i, std::size_t a_i, std::size_t a_j, float dx2, float weight, const std::vector<int> &mol_id_,
+  const std::vector<int> &natmol2_, const std::vector<float> &density_bins_,
+  const std::vector<float> &inv_num_mol_, std::vector<std::vector<std::mutex>> &frame_same_mutex_, 
+  std::vector<std::vector<std::vector<std::vector<float>>>> &intram_mat_density_
 )
 {
   std::size_t same_mutex_index = cmdata::indexing::mutex_access(mol_id_[i], a_i, a_j, natmol2_);
@@ -46,13 +46,13 @@ void intra_mol_routine(
 }
 
 void inter_mol_same_routine(
-  int i, std::size_t mol_i, std::size_t a_i, std::size_t a_j, double dx2, double weight, 
-  const std::vector<int> &mol_id_, const std::vector<int> &natmol2_, const std::vector<double> &density_bins_,
-  std::vector<std::vector<std::mutex>> &frame_same_mutex_, std::vector<std::vector<double>> &frame_same_mat_,
-  std::vector<std::vector<std::vector<std::vector<double>>>> &interm_same_mat_density_
+  int i, std::size_t mol_i, std::size_t a_i, std::size_t a_j, float dx2, float weight, 
+  const std::vector<int> &mol_id_, const std::vector<int> &natmol2_, const std::vector<float> &density_bins_,
+  std::vector<std::vector<std::mutex>> &frame_same_mutex_, std::vector<std::vector<float>> &frame_same_mat_,
+  std::vector<std::vector<std::vector<std::vector<float>>>> &interm_same_mat_density_
 )
 {
-  double dist = std::sqrt(dx2);
+  float dist = std::sqrt(dx2);
   std::size_t same_access_index = cmdata::indexing::offset_same(mol_id_[i], mol_i, a_i, a_j, natmol2_);
   std::size_t same_mutex_index = cmdata::indexing::mutex_access(mol_id_[i], a_i, a_j, natmol2_);
   std::unique_lock lock(frame_same_mutex_[mol_id_[i]][same_mutex_index]);
@@ -62,13 +62,13 @@ void inter_mol_same_routine(
 }
 
 void inter_mol_cross_routine(
-  int i, int j, std::size_t mol_i, std::size_t mol_j, std::size_t a_i, std::size_t a_j, double dx2, double weight,
+  int i, int j, std::size_t mol_i, std::size_t mol_j, std::size_t a_i, std::size_t a_j, float dx2, float weight,
   const std::vector<int> &mol_id_, const std::vector<int> &natmol2_, const std::vector<std::vector<int>> &cross_index_,
-  const std::vector<double> &density_bins_, std::vector<std::vector<std::mutex>> &frame_cross_mutex_,
-  std::vector<std::vector<double>> &frame_cross_mat_, std::vector<std::vector<std::vector<std::vector<double>>>> &interm_cross_mat_density_
+  const std::vector<float> &density_bins_, std::vector<std::vector<std::mutex>> &frame_cross_mutex_,
+  std::vector<std::vector<float>> &frame_cross_mat_, std::vector<std::vector<std::vector<std::vector<float>>>> &interm_cross_mat_density_
 )
 {
-  double dist = std::sqrt(dx2);
+  float dist = std::sqrt(dx2);
   std::size_t cross_access_index = cmdata::indexing::offset_cross(mol_id_[i], mol_id_[j], mol_i, mol_j, a_i, a_j, natmol2_);
   std::size_t cross_mutex_index = cmdata::indexing::mutex_access(mol_id_[j], a_i, a_j, natmol2_);
   std::unique_lock lock(frame_cross_mutex_[cross_index_[mol_id_[i]][mol_id_[j]]][cross_mutex_index]);
@@ -78,8 +78,8 @@ void inter_mol_cross_routine(
 }
 
 void normalize_histo(
-    std::size_t i, int ii, int jj, double norm, double inv_num_mol_same,
-    std::vector<std::vector<std::vector<std::vector<double>>>> &data
+    std::size_t i, int ii, int jj, float norm, float inv_num_mol_same,
+    std::vector<std::vector<std::vector<std::vector<float>>>> &data
 )
 {
   std::transform(
